requirements.txt
#Contains the downloadables required to run the app
streamlit
openai
python-dotenv
wordfreq